BE - feat(gemini): add Gemini API client and response model for listing available AI models; refactor credit management in ApiKey and User entities to support new credit system; enhance credit transaction handling in services and repositories for improved accuracy and atomic operations

+++ backend/src/main/java/com/nta/domain/creditTransaction/Service.java
@@ -25,7 +25,7 @@ public class Service {
 
     @Transactional
     public CreditTransaction reserveCredit(Long userId, Long amount) {
-        int updated = userRepository.reserveCredits(userId, (long) amount);
+        int updated = userRepository.reserveCredits(userId, amount.intValue());
 
         if (updated == 0) {
             throw new AppException(ErrorCode.NOT_ENOUGH_CREDITS);
@@ -60,7 +60,7 @@ public class Service {
             throw new AppException(ErrorCode.UNAUTHORIZED);
         }
 
-        userRepository.addCredits(ownerId, Math.abs(tx.getAmount()));
+        userRepository.addCredits(ownerId, Math.abs(tx.getAmount().intValue()));
 
         tx.setStatus(TransactionStatus.FAILED);
     }
