+++ backend/src/main/java/com/nta/domain/user/dto/response/CreditBalanceResponse.java
@@ -14,5 +14,5 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class CreditBalanceResponse {
 
-    Long credits;
+    Integer credits;
 }
